requirements.txt
# Solana Projects Auto-Updater

## Python Dependencies
requests>=2.31.0
pyyaml>=6.0.1

## Development Dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
